requirements.txt
torch
timm==0.6.7
ultralytics
opencv-python
pillow